feat: :triangular_flag_on_post: write routing structure
