+++ src/services/room.service.ts
@@ -1,0 +1,16 @@
+import { ObjectId } from "mongodb";
+
+// import { connectToMongoDB, closeMongoDBConnection } from "../configs/db.config";
+
+const getRoomById = async (id: ObjectId) => {
+  return "";
+};
+
+const createRoom = async () => {
+  return "";
+};
+
+export default {
+  getRoomById,
+  createRoom,
+};
